test(users): split TestUpdateUserRoles into test_users_roles.py (file-size limit)
